fix: language switcher issues
